Move pair related functions into `pair.hpp` from `list.hpp`
