+++ include/meevax/kernel/pair.hpp
@@ -175,6 +175,82 @@ inline namespace kernel
   };
 
   auto operator <<(std::ostream &, pair const&) -> std::ostream &;
+
+  template <typename T, typename U, REQUIRES(std::is_constructible<pair, T, U>)>
+  auto operator |(T&& x, U&& y) -> decltype(auto)
+  {
+    return make<pair>(std::forward<decltype(x)>(x),
+                      std::forward<decltype(y)>(y));
+  }
+
+  inline auto cons = [](auto&&... xs) constexpr
+  {
+    return (std::forward<decltype(xs)>(xs) | ...);
+  };
+
+  inline auto xcons = [](auto&& x, auto&& y) constexpr
+  {
+    return cons(std::forward<decltype(y)>(y),
+                std::forward<decltype(x)>(x));
+  };
+
+  template <auto N, typename T>
+  auto get(T&& x) -> decltype(auto)
+  {
+    if constexpr (std::is_same_v<std::decay_t<T>, pair::iterator>)
+    {
+      return std::get<N>(*x.current);
+    }
+    else if constexpr (std::is_same_v<std::decay_t<T>, object>)
+    {
+      return std::get<N>(x.template as<pair>());
+    }
+    else
+    {
+      return std::get<N>(x);
+    }
+  }
+
+  inline auto car = [](auto&& x) -> decltype(auto)
+  {
+    return get<0>(std::forward<decltype(x)>(x));
+  };
+
+  inline auto cdr = [](auto&& x) -> decltype(auto)
+  {
+    return get<1>(std::forward<decltype(x)>(x));
+  };
+
+  inline constexpr auto caar = compose(car, car);
+  inline constexpr auto cadr = compose(car, cdr);
+  inline constexpr auto cdar = compose(cdr, car);
+  inline constexpr auto cddr = compose(cdr, cdr);
+
+  inline constexpr auto caaar = compose(car, caar);
+  inline constexpr auto caadr = compose(car, cadr);
+  inline constexpr auto cadar = compose(car, cdar);
+  inline constexpr auto caddr = compose(car, cddr);
+  inline constexpr auto cdaar = compose(cdr, caar);
+  inline constexpr auto cdadr = compose(cdr, cadr);
+  inline constexpr auto cddar = compose(cdr, cdar);
+  inline constexpr auto cdddr = compose(cdr, cddr);
+
+  inline constexpr auto caaaar = compose(car, caaar);
+  inline constexpr auto caaadr = compose(car, caadr);
+  inline constexpr auto caadar = compose(car, cadar);
+  inline constexpr auto caaddr = compose(car, caddr);
+  inline constexpr auto cadaar = compose(car, cdaar);
+  inline constexpr auto cadadr = compose(car, cdadr);
+  inline constexpr auto caddar = compose(car, cddar);
+  inline constexpr auto cadddr = compose(car, cdddr);
+  inline constexpr auto cdaaar = compose(cdr, caaar);
+  inline constexpr auto cdaadr = compose(cdr, caadr);
+  inline constexpr auto cdadar = compose(cdr, cadar);
+  inline constexpr auto cdaddr = compose(cdr, caddr);
+  inline constexpr auto cddaar = compose(cdr, cdaar);
+  inline constexpr auto cddadr = compose(cdr, cdadr);
+  inline constexpr auto cdddar = compose(cdr, cddar);
+  inline constexpr auto cddddr = compose(cdr, cdddr);
 } // namespace kernel
 } // namespace meevax
